fix for intermitent empty jenkins file

+++ git_commit.sh
@@ -7,15 +7,23 @@ do
 
   git checkout -b perf-test-$b
   parallel=$(($b+4))
-  awk -v p=$parallel 'BEGIN {FS="]"; OFS="";} NR==1 {$1=$1", "p"]"} {print $0}' JenkinsFile | tee JenkinsFile
-        while ! [ -s JenkinsFile ]; do
-                git stash
-		git clean -df
-                git checkout perf-test-$(($b-1))
-                git branch -D perf-test-$b
-                git checkout -b perf-test-$b
-                awk -v p=$parallel 'BEGIN {FS="]"; OFS="";} NR==1 {$1=$1", "p"]"} {print $0}' JenkinsFile >JenkinsFile >$(tty)
-        done
+  awk -v p=$parallel 'BEGIN {FS="]"; OFS="";} NR==1 {$1=$1", "p"]"} {print $0}' Jenkinsfile | tee Jenkinsfile
+        while true
+	do
+		if [ -s Jenkinsfile ]
+		then
+			break
+		else
+                	git stash
+			git stash drop
+			git clean -df
+                	git checkout perf-test-$(($b-1))
+                	git branch -D perf-test-$b
+                	git checkout -b perf-test-$b
+                	awk -v p=$parallel 'BEGIN {FS="]"; OFS="";} NR==1 {$1=$1", "p"]"} {print $0}' Jenkinsfile >Jenkinsfile
+			cat JenkinsFile
+        	fi
+	done
   git status
   git add .
   git status
@@ -23,7 +31,7 @@ do
   git push origin perf-test-$b
 # commit frequency is calculated based on the time interval i.e. sleep
   sleep 6
-  echo $b >> ~/JaaS/iteration-output-test.txt
+  echo $b >> ~/JaaS/iteration-output-test1.txt
   b=$(($b+1))
 
 done
